fix member functions in obstacle class
