+++ src/Class/Obstacle.h
@@ -1,7 +1,6 @@
 #ifndef OBSTACLE_H
 #define OBSTACLE_H
 
-#define CELL_IN_LANE 13
 #include <cstdlib>
 #include <vector>
 
@@ -14,7 +13,6 @@ public:
     Obstacle();
     ObstacleName randomObstacleName();
     int randomCoordinateXOfObstacles();
-    Obstacle createObstacles();
 
 private:
     int coordinateXOfObstacles;
